Move FunctionSignature creation to display
